Updated structure, fixed context loading

+++ src/main/java/Sports_Manager/TeamManager/repos/VenueRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VenueRepo extends JpaRepository<Venue,Integer> {
+public interface VenueRepo extends JpaRepository<Venue,Long> {
 }
